moved open/close modal logic into its own component
